refactor: move cross-context adapters to composition root, SourceStat to domain
- Move LeadCheckerAdapter from prospects/ to cmd/server/adapters.go
- Move ProspectRepoAdapter from inbox/ to cmd/server/adapters.go
- Move SourceStat and StatsReader from usecase to domain/ports.go
- No cross-context imports in business packages
